Refactor TasksService to include token payload in create, update, and delete methods for user authorization

+++ src/tasks/tasks.service.ts
@@ -3,6 +3,7 @@ import { CreateTaskDto } from './dto/create-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { PaginationDto } from 'src/common/dto/pagination.dto';
+import { PayloadTokenDto } from 'src/auth/dto/payload-token.dto';
 
 @Injectable()
 export class TasksService {
@@ -52,13 +53,13 @@ export class TasksService {
         }
     }
 
-    async create(createTaskDto: CreateTaskDto) {
+    async create(createTaskDto: CreateTaskDto, tokenPayload: PayloadTokenDto) {
         try {
             const task = await this.prisma.task.create({
                 data: {
                     name: createTaskDto.name,
                     description: createTaskDto.description,
-                    userId: createTaskDto.userId,
+                    userId: tokenPayload.sub,
                     completedAt: false,
                 }
             });
@@ -68,7 +69,7 @@ export class TasksService {
         }
     }
 
-    async update(id: number, updateTaskDto: UpdateTaskDto) {
+    async update(id: number, updateTaskDto: UpdateTaskDto, tokenPayload: PayloadTokenDto) {
         try {
             const findTask = await this.prisma.task.findFirst({
                 where: { id },
@@ -76,6 +77,10 @@ export class TasksService {
 
             if (!findTask) {
                 throw new HttpException('Essa tarefa não existe!', HttpStatus.NOT_FOUND);
+            }
+
+            if (findTask.userId !== tokenPayload.sub) {
+                throw new HttpException('Você não tem permissão para editar essa tarefa!', HttpStatus.BAD_REQUEST);
             }
 
             const task = await this.prisma.task.update({
@@ -94,7 +99,7 @@ export class TasksService {
         }
     }
 
-    async delete(id: number) {
+    async delete(id: number, tokenPayload: PayloadTokenDto) {
         try {
             const findTask = await this.prisma.task.findFirst({
                 where: { id },
@@ -102,6 +107,10 @@ export class TasksService {
 
             if (!findTask) {
                 throw new HttpException('Essa tarefa não existe!', HttpStatus.NOT_FOUND);
+            }
+
+            if (findTask.userId !== tokenPayload.sub) {
+                throw new HttpException('Você não tem permissão para deletar essa tarefa!', HttpStatus.BAD_REQUEST);
             }
 
             await this.prisma.task.delete({
